Simplified structure of the Lead class.
Originally I'd added a public member for each field of the Lead Entry,
to try and make it easy to cast unserialzed JSON into these objects.

However it's not possible to do that, you can't do (Lead) $stdClass so I
mived the unserialized JSON object into a private member & added a new
base class "Entry" which takes the serialized JSON on it's constructor &
provides toJSON & toArray methods for use elsewhere (like the SugarAPI
class, which uses toArray)

I've kept the Lead class though as it nicely encapsulates what the
object is, and I can see the need for per Entry type validation at a
later date.